Instagram

+++ services/upload_to_insta/upload_bot.py
@@ -27,7 +27,8 @@ class InstaBot:
             file_path = os.path.join(folder, filename)
             try:
                 if os.path.isfile(file_path) or os.path.islink(file_path):
-                    os.unlink(file_path)
+                    if(not file_path.endswith('needed')):
+                        os.unlink(file_path)
                 elif os.path.isdir(file_path):
                     shutil.rmtree(file_path)
             except Exception as e:
